+++ Source/Config.cs
@@ -7,6 +7,7 @@ namespace LCWildCardMod.Config
 {
     public class WildCardConfig
     {
+        BepInEx.Logging.ManualLogSource log = WildCardMod.Log;
         internal readonly List<ConfigEntry<bool>> isScrapEnabled = new List<ConfigEntry<bool>>();
         internal readonly List<ConfigEntry<string>> scrapSpawnWeights = new List<ConfigEntry<string>>();
         internal readonly List<ConfigEntry<bool>> isSkinEnabled = new List<ConfigEntry<bool>>();
@@ -16,25 +17,26 @@ namespace LCWildCardMod.Config
         internal WildCardConfig(ConfigFile cfg, List<Item> scrapList, List<Skin> skinList)
         {
             cfg.SaveOnConfigSet = false;
-            foreach (Item scrapListItem in scrapList)
+            for (int i = 0; i < scrapList.Count; i++)
             {
-                bool defaultEnabled = scrapListItem.spawnPrefab.GetComponent<AdditionalInfo>().defaultEnabled;
-                string defaultRarities = scrapListItem.spawnPrefab.GetComponent<AdditionalInfo>().defaultRarities;
-                bool isBonus = scrapListItem.spawnPrefab.GetComponent<AdditionalInfo>().isBonus;
+                bool defaultEnabled = scrapList[i].spawnPrefab.GetComponent<AdditionalInfo>().defaultEnabled;
+                string defaultRarities = scrapList[i].spawnPrefab.GetComponent<AdditionalInfo>().defaultRarities;
+                bool isBonus = scrapList[i].spawnPrefab.GetComponent<AdditionalInfo>().isBonus;
                 if (isBonus)
                 {
-                    bonusString = $"{bonusString}, {scrapListItem.itemName}";
+                    bonusString = $"{bonusString}, {scrapList[i].itemName}";
                 }
-                isScrapEnabled.Add(cfg.Bind("Scrap", $"Enable {scrapListItem.itemName}?", defaultEnabled, ""));
-                scrapSpawnWeights.Add(cfg.Bind("Scrap", $"Spawn Weights of {scrapListItem.itemName}!", defaultRarities, "For example: All:20,Vanilla:20,Modded:20,Experimentation:20"));
+                isScrapEnabled.Add(cfg.Bind("Scrap", $"Enable {scrapList[i].itemName}?", defaultEnabled, ""));
+                scrapSpawnWeights.Add(cfg.Bind("Scrap", $"Spawn Weights of {scrapList[i].itemName}!", defaultRarities, "For example: All:20,Vanilla:20,Modded:20,Experimentation:20"));
+
             }
             assortedScrap = cfg.Bind("Scrap", "Enable/Disable supplementary scrap overall", true, bonusString);
-            foreach (Skin skinListSkin in skinList)
+            for (int i = 0; i < skinList.Count; i++)
             {
-                bool defaultEnabled = skinListSkin.skinEnabled;
-                int defaultChance = skinListSkin.skinChance;
-                isSkinEnabled.Add(cfg.Bind("Skins", $"Enable {skinListSkin.skinName}?", defaultEnabled, ""));
-                skinApplyChance.Add(cfg.Bind("Skins", $"Chance that {skinListSkin.targetEnemy.enemyName} will spawn as {skinListSkin.skinName}", defaultChance, "Percentage chance, 0 - 100"));
+                bool defaultEnabled = skinList[i].skinEnabled;
+                int defaultChance = skinList[i].skinChance;
+                isSkinEnabled.Add(cfg.Bind("Skins", $"Enable {skinList[i].skinName}?", defaultEnabled, ""));
+                skinApplyChance.Add(cfg.Bind("Skins", $"Weighted chance that {skinList[i].targetEnemy.enemyName} will spawn as {skinList[i].skinName}", defaultChance, "Must be an integer greater than, or equal to, 0"));
             }
             ClearOrphanedEntries(cfg);
             cfg.Save();