+++ Source/Items/Fyrus/FyrusAttach.cs
@@ -1,5 +1,4 @@
 ﻿using UnityEngine;
-
 namespace LCWildCardMod.Items.Fyrus
 {
     public class FyrusAttach : MonoBehaviour